fixing some TODO items
